// src/payload/payload-types.ts
export interface Media {
  id: string;
  alt: string;
  caption?:
    | {
        [k: string]: unknown;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
}
// ...
export interface Movie {
  id: string;
  title: string;
  release_date: string;
  actors?: (string | Actor)[] | null;
  poster: string | Media;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "actors".
 */
export interface Actor {
  id: string;
  first_name?: string | null;
  last_name?: string | null;
  full_name?: string | null;
}